% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryTargetViz.R
\name{BinaryTargetViz}
\alias{BinaryTargetViz}
\title{BinaryTargetViz
}
\usage{
BinaryTargetViz(var)
}
\value{
A graph
}
\description{
BinaryTargetViz
Visualization code in shiny app.
It takes the dataframes from current memory and visualize the graph according to the selected columns
}
\examples{
x <- sapply(sapply(ls(), get), is.data.frame)
m <- data.frame(names(x)[(x == TRUE)])
BinaryTargetViz(m)
}
\author{
Jishnu S
}
